% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexml.R
\name{add_provenance_record}
\alias{add_provenance_record}
\title{Generate and add provenance record to nexml}
\usage{
add_provenance_record(nexml, cmd = NA, creator = Sys.getenv("USER"))
}
\arguments{
\item{nexml}{the \link[RNeXML:nexml]{nexml} object to which to add provenance
documentation}

\item{cmd}{character, the command (such as a function invocation) to document
in the provenance record. If \code{NA} (the default), the invocation of the
function calling this one will be used as the command.}

\item{creator}{character, a value identifying the person running the
software. The default is the system's \code{USER} environment variable.}
}
\value{
A \link[RNeXML:nexml]{nexml} object with provenance records added.
}
\description{
If the content of a \link[RNeXML:nexml]{nexml} object is modified, this
function can add a provenance record documenting the modification.
The provenance documentation is added to the toplevel metadata (i.e.,
at the \link[RNeXML:get_metadata]{"nexml" level}).
}
\details{
At present, for each invocation this implementation will do the following:
\enumerate{
\item If the top-level metadata for the nexml object contains
\code{dc:description} annotations(s), they are moved to being nested
within a \code{dcterms:provenance} annotation, and prefixed with
"Original description:".
\item A \code{dcterms:provenance} annotation is added, with nested properties
\code{dc:creator} (see parameter \code{creator}), \code{dcterms:modified} (current
time), and \code{dc:description}. The latter gives the command to document,
see parameter \code{cmd}.
\item A provenance record using \href{https://www.w3.org/TR/prov-o/}{Prov-O}
(a W3C recommendation) nested within \code{prov:wasGeneratedBy} is added.
In RDF Turtle representation, the record would have the following structure
(cf. \href{http://purl.obolibrary.org/obo/IAO_0000591}{obo:IAO_0000591}):\preformatted{:nexml prov:wasGeneratedBy [
  prov:endedAtTime "2019-06-20 15:09:08 GMT" ;
  prov:wasAssociatedWith [
    a obo:IAO_0000591 ;
    dc:title "rphenoscape" ;
    dcterms:hasVersion "<rphenoscape version>" ;
  ] ;
  prov:wasAssociatedWith [
    a prov:Person ;
    foaf:name "<creator>" ;
  ] ;
  prov:used [
    prov:value "<modifying command>" ;
  ] ;
] .
}
}
}
