% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_get_IRI.R
\name{anatomy_ontology_iris}
\alias{anatomy_ontology_iris}
\title{Get IRIs of ontologies with anatomy terms}
\usage{
anatomy_ontology_iris()
}
\value{
A character vector
}
\description{
Obtains the IRIs of anatomical entity ontologies in the Phenoscape KB.
Note that this will normally include the Gene Ontology (GO), because it
contains a substantial number of anatomical entity terms in its subcellular
location branch.
}
