% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-weights.R
\name{corpus_size}
\alias{corpus_size}
\title{Obtain the size of different corpora}
\usage{
corpus_size(
  corpus = c("taxon_annotations", "taxon-variation", "annotated-taxa",
    "gene_annotations", "genes", "states")
)
}
\arguments{
\item{corpus}{the name of the corpus, currently one of the following:
\itemize{
\item "taxon_annotations" (counts phenotype annotations to character states and thus taxa),
\item "states" (counts character states),
\item "taxon-variation" (counts taxa among variation profiles),
\item "annotated-taxa" (counts taxa with annotations),
\item "gene_annotations" (counts phenotype annotations to genes), and
\item "genes" (counts genes)
}

Note that at present "gene_annotations" is not yet supported by the Phenoscape API.
Unambiguous abbreviations of corpus names are acceptable.

Note that previously "taxa" was allowed as a corpus, but is no longer supported.
The "taxon-variation" corpus is the equivalent of the deprecated "taxa" corpus.}
}
\value{
The total size of the specified corpus as an integer number.
}
\description{
Obtains the size of a certain number of predefined corpora. The total size
of a corpus is important for calculating term frequencies. That is, for a given
corpus, the possible range for any term frequency is between 0 and the corpus size.
}
\details{
Corpus sizes are cached per session after they have first been obtained.
Thus, if the Phenoscape KB changes, a session needs to be restarted to
have those changes be reflected.
}
\examples{
corpus_size("taxon-variation")
corpus_size("annotated-taxa")
corpus_size("states")
corpus_size("genes")
}
