% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexml.R
\name{get_char_matrix_meta}
\alias{get_char_matrix_meta}
\title{Obtains taxa and entity metadata from a nexml object}
\usage{
get_char_matrix_meta(nex)
}
\arguments{
\item{nex, }{a [nexml][\link[RNeXML:nexml-class]{RNeXML::nexml} object}
}
\value{
list: A list containing two data frames. The first list item (id_taxa) contains
a data frame with columns label, href, out and otus columns. The second list item
(id_entities) contains a data frame with columns label, href, and char.
}
\description{
Obtains taxa and entity metadata from a nexml object
}
\examples{
\dontrun{
nex <- get_ontotrace_data(taxon = c("Ictalurus", "Ameiurus"), entity = "fin spine")
get_char_matrix_meta(nex)

slist <- get_studies(taxon = "Ictalurus australis", entity = "fin spine")
nex_list <- get_study_data(slist$id)
nex <- nex_list[[1]]
get_char_matrix_meta(nex)
}
}
