% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkb_get.R
\name{get_data}
\alias{get_data}
\alias{get_json_data}
\alias{pk_GET}
\alias{get_csv_data}
\alias{get_nexml_data}
\title{Get data from an API endpoint}
\usage{
get_json_data(
  url,
  query,
  verbose = FALSE,
  ensureNames = NULL,
  forceGET = FALSE,
  cleanNames = FALSE
)

get_csv_data(url, query, ..., verbose = FALSE, forceGET = FALSE)

get_nexml_data(url, query, verbose = FALSE, forceGET = FALSE)
}
\arguments{
\item{url}{character, the URL of the API endpoint to query}

\item{query}{list, the query parameters and values in the form of a named list}

\item{verbose}{logical, whether to print message when parsing results}

\item{ensureNames}{character, which column or list names to ensure are included
in the result to be returned. If result returned by the API endpoint does
not include them, they will be added, with NA values.}

\item{forceGET}{logical, whether to force using the HTTP GET method for API
access regardless of request length. The default is FALSE, meaning queries
exceeding a certain size for transmission will automatically use HTTP POST
for accessing the KB API.}

\item{cleanNames}{for \code{get_json_data}, logical, whether to clean names in the returned data
replacing  "@id" with "id"}

\item{...}{for \code{get_csv_data}, additional parameters to be passed on to
\link[utils:read.table]{read.csv()}}
}
\value{
For \code{get_json_data}, a data frame or list, depending on the result of
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.

For \code{get_csv_data}, a data frame.

For \code{get_nexml_data}, a \link[RNeXML:nexml-class]{nexml} object.
}
\description{
Obtains and parses data from JSON, CSV, and NeXML-returning API endpoints,
respectively.
}
\details{
These are package-internal functions.
}
