% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBinfo.R
\name{get_KBinfo}
\alias{get_KBinfo}
\alias{print.KBinfo}
\title{Summary and metadata about the Phenoscape KB}
\usage{
get_KBinfo()

\method{print}{KBinfo}(x, ..., tz = "")
}
\arguments{
\item{x}{the object of class "KBinfo" to be printed}

\item{...}{additional paramaters that might be passed to print methods,
ignored here}

\item{tz}{the timezone in which to print timestamp values, defaults to
local timezone}
}
\value{
A list of class "KBinfo" with summary statistics of annotation counts
and other KB metadata (specifically, a timestamp for the current KB build).
}
\description{
Returns the date of the current release of the Phenoscape Knowledgebase (KB) and
counts of annotated matrices, taxa, phenotypes, characters, and states.
}
\examples{
kbmeta <- get_KBinfo()
names(kbmeta)
kbmeta
}
