% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_study.R
\name{get_studies}
\alias{get_studies}
\title{Query the list of studies by taxa, anatomical entities, and qualities.}
\usage{
get_studies(taxon = NA, entity = NA, quality = NA,
  phenotype = NA, includeRels = c("part of"), relation = c("part of"))
}
\arguments{
\item{taxon}{character. The taxon by which to filter, if any.}

\item{entity}{character. The anatomical entity by which to filter, if any.}

\item{quality}{character. The phenotypic quality by which to filter, if any.}

\item{phenotype}{character. The phenotype (as its identifier) by which to filter, if any.
If provided, matching studies (through its one or more of its character states)
must be linked to the given phenotype, or one subsumed by it. This must be provided
as identifier, no text search will be performed for resolution.}

\item{includeRels}{character or vector of characters. The names of relationships
for anatomical entities to include in addition to subtype (\code{is_a}, \code{rdfs:subClassOf}).
Defaults to \code{"part_of"}. Set to \code{FALSE} to not include any additional relationships.
Otherwise one or more of \code{"part of"}, \code{"historical homologous to"}, and
\code{"serially homologous to"}, or set to \code{TRUE} to include all possible ones. It is
acceptable to use unambiguous prefixes, for example \code{"historical homolog"}.}

\item{relation}{character. Deprecated, for backwards compatibility defaults to
\code{part of}. Only used if \code{includeRels} is left at its default value.}
}
\value{
A data.frame with columns "id" and "label", or an empty list if no
matching study was found.
}
\description{
Return studies that contain taxa which are members of the optional input taxon,
and characters which have phenotype annotations subsumed by the given entity and quality
terms.
}
\examples{
\dontrun{
# by default, parts are included
slist <- get_studies(taxon = "Siluridae", entity = "fin")
colnames(slist)
nrow(slist)

# can also disable parts
slist <- get_studies(taxon = "Siluridae", entity = "fin", includeRels = FALSE)
nrow(slist)

# or filter studies only by entity, including their parts
slist <- get_studies(entity = "pelvic fin", includeRels = c("part of"))
nrow(slist)

# or filter studies only by entity, including their parts
slist <- get_studies(entity = "pelvic fin", includeRels = c("part of"))
nrow(slist)

# including not only parts but also historical and serial homologs
slist <- get_studies(entity = "pelvic fin",
                     includeRels = c("part of",
                                     "serially homologous to",
                                     "historical homologous to"))
nrow(slist)
# relationship names can be given as prefixes
slist1 <- get_studies(entity = "pelvic fin",
                      includeRels = c("part", "serial", "historical"))
nrow(slist1) == nrow(slist)

# or apply no filter, obtaining all studies in the KB
slist <- get_studies()
nrow(slist)
}
}
