% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_terms.R
\name{get_term_label}
\alias{get_term_label}
\title{Obtains the labels for a list of terms}
\usage{
get_term_label(term_iris, preserveOrder = FALSE, verbose = FALSE)
}
\arguments{
\item{term_iris}{character, a list of term IRIs}

\item{preserveOrder}{logical, whether the resulting data frame (see Value)
is to be in the same row order as \code{termIRIs}. The default is not to preserve
order.}

\item{verbose}{logical, whether to print information about possibly
time-consuming operations.}
}
\value{
A data.frame with columns "id" and "label".
}
\description{
Attempts to obtain the label for each term, identified by IRI, in the input
list. Terms for which no label is found in the database will have NA as the
label in the result (see Value).
}
