% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{memoise}
\alias{memoise}
\title{Memoise a function optionally with persistent cache}
\usage{
memoise(f, persistName = NULL)
}
\arguments{
\item{f}{the function to \link[memoise:memoise]{memoise}}

\item{persistName}{character, a name for the persistent cache to load. Do not
include extension or path. The corresponding file is assumed to be part
of the installed package. If there is no such persistent cache file, this
is ignored on startup/load.}
}
\description{
This is a thin layer over \code{\link[memoise:memoise]{memoise::memoise()}}, by applying a few defaults.
Specifically, by default the cache is in memory, but optionally a name for
a persistent cache can be given, in which case the cache is still held (and
grown) in memory, but is initialized from a file if it exists.
}
\details{
The name of the persistent cache is turned into a file path by prefixing it
with the path to the \code{data-cache/} directory within the installed package,
and appending the extension \code{.rds}. Normally this location will not
be writable, i.e., the default persistent file cache will normally come with
installation, and stay unchanged.
}
\note{
This function is a shortcut meant to be used only by the package's developers.
The described behavior when providing a persistent cache name can be achieved
with files located anywhere by using \code{\link[memoise:memoise]{memoise::memoise()}} and passing
\code{\link[=cache_serializableMemory]{cache_serializableMemory()}} for the \code{cache} parameter.
}
\keyword{internal}
