% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_exclusivity.R
\name{mutually_exclusive}
\alias{mutually_exclusive}
\title{Determine mutual exclusivity between two or more phenotypes}
\usage{
mutually_exclusive(
  phenotypes,
  studies = NULL,
  progress_bar = FALSE,
  quality_opposites = NULL
)
}
\arguments{
\item{phenotypes}{character or phenotype, a vector of phenotype IDs, or
phenotype objects for which one wants to determine the mutual exclusivity.}

\item{studies}{character, a vector of study IDs (optional) if one is
interested in determining mutual exclusivity only based on the evidence from a
particular set of studies. By default \emph{all} studies in the KB are considered.}

\item{progress_bar}{logical, whether to print progress of the function.

WARNING: setting progress_bar to TRUE clears the R console by executing the
cat('\014') command before printing the progress.}

\item{quality_opposites}{dataframe, an optional dataframe containing columns
"quality.a" and "quality.b" to denote pairs of phenotypic quality terms, in
the form of their term IRIs, to be considered opposites of each others. If
provided, two phenotypes will be determined to have \emph{strong exclusivity} if
their qualities match a pair of opposites. The test will only be applied to
pairs of phenotypes in which the EQ expressions of both comprise of the same
number of entities and only a single quality term. Columns included in the
dataframe other than "quality.a" and "quality.b" will be ignored.}
}
\value{
A list consisting a matrix and a dataframe that contain mutual exclusivity
results for the phenotypes.

The mutual exclusivity type between two phenotypes is encoded as an ordered factor with the following
levels (in this order):
\itemize{
\item 'strong_compatibility'  : the data show strong evidence of mutual compatibility;
\item 'weak_compatibility'    : the data suggest mutual compatibility, but the evidence is weak;
\item 'inconclusive_evidence' : the default for phenotypes that lack links to character states;
\item 'weak_exclusivity'      : the data suggest mutual exclusivity, but the evidence is weak;
\item 'strong_exclusivity'    : the data show strong evidence of mutual exclusivity.
}

The first item in the list is a square matrix with each dimension equal to the
number of phenotypes. The cells (\emph{i},\emph{j}) of the matrix are the integer representation of
the factor level for the mutual exclusivity type between the \emph{i} th and \emph{j} th phenotype
in the list of phenotypes provided as input. The matrix is symmetric.

The second item in the list is a dataframe with five columns, the labels and ids
of each phenotype of a pair, and the mutual exclusivity type of the pair is in the
fifth column.
}
\description{
Determines whether the data in the KB includes evidence for mutual exclusivity
between all distinct pairs of phenotypes in the provided list.
}
\details{
Evidence for mutual exclusivity or compatibility between a pair of phenotypes
is taken from how they are or aren't linked to the same taxon, character, and
character state. If two phenotypes are found to be linked to the same character
state at least once, it is considered \emph{strong evidence for mutual compatibility}.
If they are found to be linked to the same character at least once, but not
to the same character state (implying they must be linked to different states of
the same character at least once), it is considered \emph{strong evidence for mutual exclusivity}.
Otherwise, if there is no character to which both are linked, the presence and absence
of data showing co-occurrence in taxa is used as \emph{weak evidence}. That is, if they are
linked to character states of different characters, but the states are both exhibited
by the same taxon at least once, it is considered as \emph{weak evidence for mutual compatibility}
(weak because the co-occurrence may be due to taxon-level polymorphism),
and if there is no such taxon, this is considered as \emph{weak evidence for mutual exclusivity}
(because lack of evidence could still be due to gaps in the data, and does therefore not
imply evidence of absence).

It is possible that for any pair of phenotypes the KB contains conflicting
evidence on mutual exclusivity. For example, two phenotypes could be linked to
the same character state for one character, and different character states for
another character (possibly from another study by another author), the first of
which would present strong evidence for mutual compatibility and the latter
strong evidence of mutual exclusivity. The current implementation does not detect
and thus ignores this possibility.
}
\examples{
# Example 1: pass phenotypes as characters with phenotype IDs
# get a specific study of interest based on which mutual exclusivity is to be determined
studies <- get_studies()
study <- studies$id[studies$label == 'Dillman et al. (2016)']

# get phenotypes ids present in the study
phenotypes <- get_phenotypes(study = study)
# select 10 phenotype ids randomly from those in the study
phenotypes_ids <- sample(phenotypes$id, 10)

# determine mutual exclusivity
exclusivity <- mutually_exclusive(phenotypes_ids)

# resultant matrix
exclusivity$matrix

# resultant dataframe
str(exclusivity$dataframe)

# filter dataframe to contain only mutually exclusive phenotypes
exclusivity_df <- exclusivity$dataframe
exclusive_phenotype_pairs <- exclusivity_df[
    exclusivity_df$mutual_exclusivity == "weak_exclusivity" |
    exclusivity_df$mutual_exclusivity == "strong_exclusivity", ]

# Example 2: pass phenotypes as phenotype objects
# sample some phenotypes objects present in the study
phenotypes_objs <- as.phenotype(sample(phenotypes$id, 4), withTaxa=TRUE)

# determine mutual exclusivity
exclusivity <- mutually_exclusive(phenotypes_objs)

# resultant matrix
exclusivity$matrix

# resultant dataframe
str(exclusivity$dataframe)

# Example 3: determine mutual exclusivity for two phenotypes
# fetch phenotypes for an entity/taxon combination
phenotypes <- get_phenotypes(entity="hyomandibular bone", taxon="Hypancistrus")

# select two phenotype ids based on the labels
phenotypes_ids <- c(
  phenotypes$id[phenotypes$label == 'hyomandibular bone in contact with prootic bone'],
  phenotypes$id[phenotypes$label == 'hyomandibular bone in contact with quadrate bone']
)

# determine mutual exclusivity
exclusivity <- mutually_exclusive(phenotypes_ids)

# exclusivity value
exclusivity$dataframe$mutual_exclusivity

# Example 4: determine mutual exclusivity for two phenotypes including opposite quality data

# create a list of phenotypes to compare (femur elongated vs femur decreased length)
phens <- get_phenotypes(entity="femur", quality="elongated")
femur_elongated_iri <- phens$id[phens$label == "femur elongated"]
phens <- get_phenotypes(entity="femur", quality="decreased length")
femur_decreased_length_iri <- phens$id[phens$label == "femur decreased length"]
phenotypes_to_compare <- c(femur_elongated_iri, femur_decreased_length_iri)

# compare the phenotypes without using opposite quality data
exclusivity <- mutually_exclusive(phenotypes_to_compare)
exclusivity$dataframe$mutual_exclusivity

# create a dataframe containing the quality opposites
elongated_iri <- find_term("elongated", matchTypes = "exact")$id
decreased_length_iri <- find_term("decreased length", matchTypes = "exact")$id
quality_opposites <- data.frame(
  quality.a = elongated_iri,
  quality.b = decreased_length_iri
)

# compare the phenotypes using opposite quality data
exclusivity <- mutually_exclusive(phenotypes_to_compare, quality_opposites=quality_opposites)
exclusivity$dataframe$mutual_exclusivity
}
