% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-type.R
\name{obo_prefix}
\alias{obo_prefix}
\title{Extract the OBO ontology prefix from IRIs}
\usage{
obo_prefix(x)
}
\arguments{
\item{x}{a list or vector of IRIs, and/or objects that have an "id" key.}
}
\value{
A character vector of the same length as the input
vector or list, with NA in the positions where extracting the OBO
ontology prefix failed.
}
\description{
Extract the OBO ontology prefix from IRIs
}
\examples{
tt <- c("http://purl.obolibrary.org/obo/UBERON_0011618",
        "http://purl.obolibrary.org/obo/PATO_0002279",
        "http://purl.obolibrary.org/obo/VTO_0071642",
        "http://purl.obolibrary.org/obo/MP_0030825",
        "http://purl.obolibrary.org/obo/NCBITaxon_7955")
obo_prefix(tt)
}
