% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{pa_dep_matrix}
\alias{pa_dep_matrix}
\title{Obtains a presence-absence dependency matrix}
\usage{
pa_dep_matrix(terms, .names = c("ID", "IRI", "label"), .labels = NULL,
  preserveOrder = FALSE, verbose = FALSE)
}
\arguments{
\item{terms}{character, the list of terms for which to compute the dependency
matrix. Can be given as term IRIs or term labels, and the list can contain
both. Terms given as labels will first be resolved to IRIs, assuming they
are from an anatomy ontology.}

\item{.names}{character, how to name the rows and columns of the resulting
matrix.
\itemize{
\item \code{"ID"} (the default): use the term IDs (the last component of the
term IRIs).
\item \code{"IRI"}: use the term IRIs as names. Note that the column names will have
\code{:}, \code{/}, and some other characters replaced with \code{.} (dot).
\item \code{"label"}: use the terms' labels (see \code{.labels} parameter).
}}

\item{.labels}{character, the labels for terms where known. Only used if
\code{.names = "label"}. If NULL (the default), labels will be looked up if \code{terms}
are provided as IRIs; elements of the \code{terms} list that are not in IRI form
are assumed to be the label. If a list, must have the same length and ordering
as \code{terms}; any NA elements will be looked up (from the corresponding term
IRI).}

\item{preserveOrder}{logical, whether to return rows (and columns) in the same
order as \code{terms}. The default is not to preserve the order.}

\item{verbose}{logical, whether to print informative messages about certain
potentially time-consuming operations.}
}
\value{
A data.fram M with M[i,j] = 1 iff the presence of term \emph{i} implies the
presence of term \emph{j}, and 0 otherwise.

The matrix will have additional attributes depending on the choice of how to
name rows and columns. If \code{.names = "ID"} (the default), the matrix will have
an attribute \code{prefixes} giving the URL prefixes removed from the term IRIs
to yield the IDs, in the order of the rows. If \code{.names = "label"}, it will
have attribute \code{term.iris}, giving the term IRIs for the rows (and columns).
Note that these extra attributes will be lost upon subsetting the returned
matrix.
}
\description{
Obtains a presence-absence dependency matrix for the given set of terms. The
resulting matrix M will have values 1 and 0, where M[i,j] = 1 iff the
presence if term \emph{i} implies the presence of term \emph{j}. Note that it follows that
M[i,j] = 1 iff the absence of term \emph{j} implies the absence of term \emph{i}.
Note also that the matrix is asymmetric; that is, M[i,j] = 1 does not imply
M[j,i] = 1 for \eqn{i\neq j}{i != j}. Terms \emph{i} and \emph{j} (\eqn{i\neq j}{i != j})
for which M[i,j] = M[j,i] = 1 are sometimes referred to as \emph{super-dependent}.
}
\examples{
\dontrun{
tl <- c("http://purl.obolibrary.org/obo/UBERON_0000981",
        "http://purl.obolibrary.org/obo/UBERON_0002103",
        "http://purl.obolibrary.org/obo/UBERON_0000976",
        "http://purl.obolibrary.org/obo/UBERON_0002102")
m <- pa_dep_matrix(tl)
m # term IDs as row and column names
id_prefixes <- attr(m, "prefixes")
id_prefixes # 4x "http://purl.obolibrary.org/obo/"

m <- pa_dep_matrix(tl, .names = "label")
m # term labels as row and column names
mat_terms <- attr(m, "term.iris")
mat_terms # term IRIs in the same order as rows (and columns)
}
}
