% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{phenotype_matches}
\alias{phenotype_matches}
\title{Which phenotypes match filter}
\usage{
phenotype_matches(x, studies)
}
\arguments{
\item{x}{character, vector of phenotype IDs to test. A data.frame is
allowed if it has a column labeled "id".}

\item{studies}{character, vector of study IDs for which to test. A data.frame
is allowed if it has a column labeled "id".}
}
\value{
a logical vector of the same length as the vector of phenotypes, with
TRUE as element if the corresponding phenotype matches, and FALSE otherwise.
}
\description{
Determines which of one or more phenotype IDs match the given filter.
}
\details{
At present, the only supported query filter is a list of studies (as their
IDs). A phenotype matches if it is linked to at least one of the studies.
}
\examples{
x <- get_phenotypes(entity = "basihyal bone")
nrow(x)
# which of these are in the same study or studies as the first one?
phenotype_matches(x, get_studies(phenotype = x$id[1]))
}
