% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_classification.R
\name{pk_class}
\alias{pk_class}
\alias{pk_taxon_class}
\title{Classificatoin}
\usage{
pk_taxon_class(x, verbose = TRUE)
}
\arguments{
\item{term}{characters}
}
\value{
A list containing data.frame
}
\description{
Return direct superclasses, direct subclasses, and equivalent classes of a given term
}

