% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_get_xml.R
\name{pk_get_ontotrace_xml}
\alias{pk_get_ontotrace_xml}
\title{Return the NeXML object}
\usage{
pk_get_ontotrace_xml(taxon, entity, relation = "part of",
  variable_only = TRUE)
}
\arguments{
\item{taxon}{character: Required. One or a vector of taxon names.}

\item{entity}{characters: Required.
A single character string or a vector of anatomical class expressions.}

\item{relation}{character string: Optional.
The relationship to the entities to be included in the result. Must be
either "part of" or "develops from", or set to NA to disable.
Default is "part of".}

\item{variable_only}{logical: Optional.
Whether to only include characters that are variable across the selected
set of taxa. Default is TRUE.}
}
\value{
RNeXML object
}
\description{
Return the NeXML object
}
\examples{
\dontrun{
nex0 <- pk_get_ontotrace_xml(taxon = "Ictalurus australis", entity = "fin")
}
\dontrun{
nex <- pk_get_ontotrace_xml(taxon = c("Ictalurus", "Ameiurus"),
                            entity = "fin spine")
}
}
