% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_study.R
\name{pk_get_study}
\alias{pk_get_study}
\title{pk_get_study}
\usage{
pk_get_study(nexmls)
}
\arguments{
\item{nexmls, }{a list of \link[RNeXML:nexml-class]{nexml} objects}
}
\value{
A list of data.frames containing matrices
}
\description{
pk_get_study
}
\examples{
\dontrun{
slist <- get_studies(taxon = "Ameiurus", entity = "pelvic splint")
nex_list <- get_study_data(slist$id) # get the list of NeXML objects for the studies
pk_get_study(nex_list) # retrieve the study matrices
pk_get_study_meta(nex_list) # retrieve the meta data for the studies
}
}
