% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_study.R
\name{pk_get_study_meta}
\alias{pk_get_study_meta}
\title{pk_get_study_meta}
\usage{
pk_get_study_meta(nexmls)
}
\arguments{
\item{nexmls, }{a list of NeXml object}
}
\description{
pk_get_study_meta
}
\examples{
\dontrun{
slist <- pk_get_study_list(taxon = "Ameiurus", entity = "pelvic splint")
nex_list <- pk_get_study_xml(slist$id) # get the list of NeXML objects for the studies
pk_get_study(nex_list) # retrieve the study matrices
pk_get_study_meta(nex_list) # retrieve the meta data for the studies
}
}

