% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_get_xml.R
\name{pk_get_study_xml}
\alias{pk_get_study_xml}
\title{pk_get_study_xml}
\usage{
pk_get_study_xml(study_ids)
}
\arguments{
\item{study_ids, }{a list of study ids.}
}
\value{
a list of nexml objects
}
\description{
pk_get_study_xml
}
\examples{
\dontrun{
slist <- pk_get_study_list(taxon = "Ameiurus", entity = "pelvic splint")
nex_list <- pk_get_study_xml(slist$id)
}
}

