% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_get_xml.R
\name{pk_get_xml}
\alias{pk_get_xml}
\alias{pk_ontotrace_xml}
\title{Returns a NeXML object.}
\usage{
pk_ontotrace_xml(taxon, entity, relation = "part of", variable_only = TRUE)
}
\arguments{
\item{taxon}{character: Required. A single character string or a vector of taxa.}

\item{entity}{characters: Required. A single character string or a vector of anatomical class expressions.}

\item{relation}{character string: Optional. Has to be either "part of" or "develops from". Default is "part of".}

\item{variable_only}{logical: Optional. Default is TRUE.}
}
\value{
RNeXML object
}
\description{
Returns a NeXML object.
}
\examples{
nex0 <- pk_ontotrace_xml(taxon = "Ictalurus", entity = "fin")
nex <- pk_ontotrace_xml(taxon = c("Ictalurus", "Ameiurus"), entity = "fin spine", get_metadata = TRUE)
}

