% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_if_descendant.R
\name{pk_if}
\alias{pk_if}
\alias{pk_if_ancestor}
\alias{pk_if_descendant}
\title{pk_if_descendant}
\usage{
pk_if_descendant(taxon, ids)

pk_if_ancestor(taxon, ids)
}
\arguments{
\item{taxon}{}

\item{ids,}{a list or vector of taxon IRIs}
}
\value{
a list of matching ids
}
\description{
pk_if_descendant
}

