% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_study.R
\name{pk_search_studies}
\alias{pk_search_studies}
\title{Query studies by taxa and anatomical entities}
\usage{
pk_search_studies(taxon, entity, relation = "part of")
}
\arguments{
\item{relation}{}
}
\value{
data.frame
}
\description{
Return studies containing taxa which are members of the optional input taxon
expression and are have annotated phenotypes which are relevant to the optional
input entity expression.
}

