% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_terms.R
\name{pk_terms}
\alias{pk_terms}
\alias{pk_taxon_detail}
\alias{pk_anatomical_detail}
\alias{pk_phenotype_detail}
\alias{pk_gene_detail}
\title{Get details about a given term}
\usage{
pk_taxon_detail(term, verbose = FALSE)

pk_anatomical_detail(term, verbose = FALSE)

pk_phenotype_detail(term, verbose = FALSE)

pk_gene_detail(term, verbose = FALSE)
}
\arguments{
\item{term}{character. The term to be searched.}

\item{verbose}{logical: optional. If TRUE (default), informative messages printed.}
}
\value{
A data.frame with term id, label, and definition
}
\description{
Retrieve details about a taxon, an anatomical structure, a gene, or a phenotype.
}
\examples{
pk_taxon_detail("Coralliozetus")
pk_anatomical_detail("basihyal bone")
pk_gene_detail("socs5")

}
