% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkb_get.R
\name{pkb_args_to_query}
\alias{pkb_args_to_query}
\title{Creates a list of named query parameters}
\usage{
pkb_args_to_query(..., includeRels = FALSE, verbose = FALSE)
}
\arguments{
\item{...}{any combination of zero or more parameters from \code{entity},
\code{quality}, \code{taxon}, and \code{study}. Any provided with value \code{NA} will be
ignored. Entity, quality, and taxon will be resolved to IRI if not
already provided as such.}

\item{includeRels}{character, in which case it is the relationship(s) for
entities to be included ("part of", "historical homologous to",
"serially homologous to"; unambiguous prefix strings are acceptable); or
logical, in which case \code{FALSE} means no relationships, and \code{TRUE} means
all available relationships. For legacy reasons, \code{NA} is treated synonymous
with \code{FALSE}. The default is \code{FALSE}.}

\item{verbose}{logical, whether to print messages when potentially time consuming
operations are run.}
}
\value{
A list of named query parameters suitatoble for several form-like
query endpoints in the Phenoscape KB API.
}
\description{
Several Phenoscape KB API endpoints use a form-like parameter list for
filtering (limiting), or, in the case of relationships, expanding, the
query result. This function aids in preparing the query string for these
endpoints. It is internal to the package.
}
