% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkb_get.R
\name{rclean_jsonld_names}
\alias{rclean_jsonld_names}
\title{Clean JSON-LD names in a list/data.frame}
\usage{
rclean_jsonld_names(x)
}
\arguments{
\item{x}{a data frame or list, typically created using \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}}
}
\value{
the input data frame or list with changed names
}
\description{
Recursively cleans names in a list/data.frame replacing \code{@id} names with \code{id}.
}
