% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_obj.R
\name{resolve_label_expression}
\alias{resolve_label_expression}
\title{Convert a label-based class expression in OWL Manchester syntax into an expression
composed of IRIs.}
\usage{
resolve_label_expression(label_expression)
}
\arguments{
\item{label_expression}{character, class expression in a label-based Manchester syntax,
e.g. "'part of' some head"}
}
\value{
A object of class "owl" with a resolved OWL Manchester expression composed of IRIs.
Note that (at least currently) these are in essence character strings, and are recognizable as such by R.
}
\description{
Convert a label-based class expression in OWL Manchester syntax into an expression
composed of IRIs.
}
