% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{restore_cache}
\alias{restore_cache}
\title{Restore a memoised function's cache from a file}
\usage{
restore_cache(f, ...)
}
\arguments{
\item{f}{the (memoised) function for which to restore the cache from file}

\item{...}{parameters to be passed on to the \code{restore()} method of the cache,
normally the file path if different from the one with which the cache was
initialized. If a file path is provided, by default a warning is issued
if it does not exist. This can be suppressed by additionally passing
\code{warnNotExists = FALSE}.}
}
\description{
Restore a memoised function's cache from a file
}
\seealso{
Other cache methods: \code{\link{cache_serializableMemory}},
  \code{\link{serialize_cache}}
}
\concept{cache methods}
