% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{serialize_cache}
\alias{serialize_cache}
\title{Serialize a memoised function's cache to disk}
\usage{
serialize_cache(f, ...)
}
\arguments{
\item{f}{the (memoised) function for which to serialize the cache to file}

\item{...}{parameters to be passed on to the \code{serialize()} method of the cache,
normally the file path if different from the one with which the cache
was initialized.}
}
\description{
Serialize a memoised function's cache to disk
}
\seealso{
Other cache methods: \code{\link{cache_serializableMemory}},
  \code{\link{restore_cache}}
}
\concept{cache methods}
