% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semsim.R
\name{tanimoto_similarity}
\alias{tanimoto_similarity}
\alias{jaccard_similarity}
\alias{cosine_similarity}
\title{Compute semantic similarity metrics between terms}
\usage{
tanimoto_similarity(subsumer_mat = NA, terms = NULL, ...)

jaccard_similarity(subsumer_mat = NA, terms = NULL, ...)

cosine_similarity(subsumer_mat = NA, terms = NULL, ...)
}
\arguments{
\item{subsumer_mat}{matrix or data.frame, the vector-encoded matrix M of
subsumers for which M[i,j] = W[i], with W[i] > 0 (weight of class \emph{i}),
if class \emph{i} subsumes term j, and 0 otherwise. A binary
(\eqn{M_{i,j} \in {0, 1}}{M\[i,j\] in {0, 1}}) encoding (i.e., W[\emph{i}] = 1)
can be obtained from \link[=subsumer_matrix]{subsumer_matrix()}.}

\item{terms}{character, optionally the list of terms (as IRIs and/or labels)
for which to generate a properly encoded subsumer matrix on the fly.}

\item{...}{parameters to be passed on to [subsumer_matrix()\link{subsumer_matrix}
if a subsumer matrix is to be generated on the fly.}
}
\value{
A matrix with M[i,j] = similarity of terms \emph{i} and \emph{j}.
}
\description{
The Tanimoto similarity ST is computed according to the definition for bit vectors
(see \href{https://en.wikipedia.org/wiki/Jaccard_index#Tanimoto's_definitions_of_similarity_and_distance}{Jaccard index at Wikipedia}).
For weights W[i] in {0, 1} it is the same as the Jaccard similarity. The
Tanimoto similarity can be computed for any term vectors, but for 1 - ST to be
a proper distance metric satisfying the triangle inequality, M[i,j] in {0, W[i]}
must hold.

The Jaccard similarity is computed using the Tanimoto similarity definition
for bit vectors
(see \href{https://en.wikipedia.org/wiki/Jaccard_index#Tanimoto's_definitions_of_similarity_and_distance}{Jaccard index at Wikipedia}).
For the results to be a valid Jaccard similarity, weights must be zero and
one. If any weights are different, a warning is issued.

The Cosine similarity \emph{SC} is computed using the Euclidean dot product formula.
See \href{https://en.wikipedia.org/wiki/Cosine_similarity#Definition}{Cosine similarity on Wikipedia}.
The metric is valid for any term vectors (columns of the subsumer matrix), i.e.,
\eqn{M_{i,j} \in {0, W_i}}{M[i,j] in {0, W[i]}} is not required. Note that
1 - \emph{SC} is not a proper distance metric, because it violates the triangle
inequality. First convert to angle to obtain a distance metric.
}
\examples{
\dontrun{
sm <- jaccard_similarity(terms = c("pelvic fin", "pectoral fin",
                                   "forelimb", "hindlimb",
                                   "dorsal fin", "caudal fin"),
                         .colnames = "label")
sm

# e.g., turn into distance matrix, cluster, and plot
plot(hclust(as.dist(1-sm)))
}
}
