% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IRI.R
\name{taxon_ontology_iris}
\alias{taxon_ontology_iris}
\title{Get IRIs of ontologies with taxonomy terms}
\usage{
taxon_ontology_iris()
}
\value{
A character vector
}
\description{
Obtains the IRIs of taxon ontologies in the Phenoscape KB.
}
