% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_classification.R
\name{term_classification}
\alias{term_classification}
\title{Get term classification}
\usage{
term_classification(x, as, verbose = FALSE)
}
\arguments{
\item{x}{character. Name of the term}

\item{as}{character. The ontology or ontologies as which to find the term.
Can be provided in several ways: (1) IRI of the ontology; (2) the ID space
for OBO ontologies such as UBERON, VTO, etc; (3) \code{"anatomy"} and \code{"taxon"}
as shorthand for all anatomy and taxon ontologies, respectively; (4) NA to
disable any filtering by defining ontology. Options (1) and (2) can be
combined. There is no default.}

\item{verbose}{logical: optional. If TRUE, prints messages prior to potentially
time-consuming operations. Default is FALSE.}
}
\value{
A list containing data.frame
}
\description{
Return direct superclasses, direct subclasses, and equivalent classes of a given term
}
