% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-weights.R
\name{term_freqs}
\alias{term_freqs}
\title{Obtains term frequencies for the Phenoscape KB}
\usage{
term_freqs(x, as = c("auto", "entity", "quality", "phenotype"),
  corpus = c("taxon_annotations", "taxa", "gene_annotations", "genes"))
}
\arguments{
\item{x}{a vector or list of one or more terms, either as IRIs or as term
objects.}

\item{as}{the category or categories of the input terms (see \code{\link[=term_category]{term_category()}}).
Supported categories are "entity", "quality", and "phenotype". The value
must either be a single category (applying to all terms), or a vector of
categories (of same length as \code{x}). If provided as "auto" (or NULL), the
category of each term is automatically determined. The default is "auto".}

\item{corpus}{the name of the corpus for which to determine frequencies.
Supported values are "taxon_annotations", "taxa", "gene_annotations", and
"genes". (At present, support for "gene_annotations" is pending support in
the Phenoscape API.) The default is "taxon_annotations".}
}
\value{
a vector of frequencies as floating point numbers (between zero
and 1.0), of the same length (and ordering) as the input list of terms.
}
\description{
Determines the frequencies for the given input list of terms, based on
the selected corpus.
}
\details{
Depending on the corpus selected, the frequencies are queried directly
from the Phenoscape API, or calculated based on query results. Currently,
the Phenoscape KB has precomputed frequencies for corpora "taxa" and
"genes".
}
\examples{
terms <- c("pectoral fin", "pelvic fin", "dorsal fin", "paired fin")
IRIs <- sapply(terms, pk_get_iri, as = "anatomy")
term_freqs(IRIs)

phens <- get_phenotypes(entity = "basihyal bone")
term_freqs(phens$id, as = "phenotype", corpus = "taxon_annotations")
term_freqs(phens$id, as = "phenotype", corpus = "taxa")

}
