% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terminfo}
\alias{terminfo}
\alias{as.terminfo}
\alias{is.terminfo}
\alias{is_valid_terminfo}
\alias{print.terminfo}
\title{Terminfo Objects}
\usage{
as.terminfo(x, ...)

is.terminfo(x)

is_valid_terminfo(x)

\method{print}{terminfo}(x, ...)
}
\arguments{
\item{x}{an object of type "terminfo" or coercible to it, or to be tested
for being of type "terminfo"}

\item{...}{additional parameters where applicable}
}
\value{
\code{as.terminfo} returns and object of type "terminfo", or a list of such objects
if the object to be coerced had multiple elements (if a vector) or rows (if a
data.frame). A terminfo object has properties "id" (ID, i.e., IRI of the
term), "label" (label of the term if one exists). If the term is a taxon
additional properties will be populated: "extinct", "rank", "common_name".
There will also be a key "classification" (a list with properties "subClassOf",
"equivalentTo" and "superClassOf").

\code{is.terminfo} returns TRUE if the object is of type "terminfo" and FALSE
otherwise.

\code{is_valid_terminfo} returns a logical vector of the same length as the input
array of objects, with TRUE for those objects in the list that are of type
"terminfo" and correspond to a term in the database.
}
\description{
\code{as.terminfo} creates an object (or a list of objects) of type "terminfo".
The object to be coerced can be a character vector (of term IRIs), or a data.frame.
In the latter case, there must be a column "id" with term IRIs.

\code{is.terminfo} tests whether an object is of type "terminfo"

\code{is_valid_terminfo} tests which of the objects in the list are valid
terminfo objects, and returns a logical vector of the same length as \code{x}.
An object is a valid terminfo object if it is of type "terminfo" and its
IRI has been found in the database.

\code{print} pretty-prints objects of type "terminfo"
}
\details{
Create and test objects of type "terminfo", and extract properties from them.
}
\examples{
# find a term iri
term_iri <- find_term('basihyal bone', matchType='exact')
# turn it into a terminfo object
obj <- as.terminfo(term_iri)
class(obj)
obj


# taxon terms have additional properties: 
term_iri <- find_term('Coralliozetus angelicus', matchType='exact')
# turn it into a terminfo object
obj <- as.terminfo(term_iri)
class(obj)
obj

}
