% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms}
\alias{terms}
\alias{taxon_info}
\alias{anatomy_term_info}
\alias{phenotypic_quality_term_info}
\alias{gene_info}
\title{Get term details (ID, label, definition)}
\usage{
taxon_info(term, verbose = FALSE)

anatomy_term_info(term, verbose = FALSE)

phenotypic_quality_term_info(term, verbose = FALSE)

gene_info(term, taxon = NA, verbose = FALSE)
}
\arguments{
\item{term}{character, the query term, either as name or IRI. Names are looked
up against taxonomies, anatomy ontologies, and PATO for \code{taxon_info},
\code{anatomy_term_info}, and \code{phenotypic_quality_term_info}, respectively.

For \code{taxon_info} this can also be a list (or vector) of terms (taxa).}

\item{verbose}{logical, whether informative messages should be printed. The
default is \code{FALSE}.}

\item{taxon}{character, the NCBI taxon name or corresponding NCBITaxon
ontology IRI for which to match the gene name.}
}
\value{
A data.frame, with at least columns "id" and "label".

For \code{taxon_info}, additional columns are "extinct" (logical),
"rank.id", "rank.label", and where available "common_name". The rows
corresponding to taxon names that failed to be resolved to IRIs will be NA.

For \code{anatomy_term_info} and \code{phenotypic_quality_term_info}, the additional
column is "definition".

For \code{gene_info}, the additional columns are "taxon.id" and "taxon.label"
for the corresponding NCBI Taxonomy ID and name, and "matchType" ('exact'
or 'partial').
}
\description{
Retrieve details about a taxon, an anatomical structure, a gene, or a phenotypic
quality.
}
\examples{
taxon_info("Coralliozetus")
anatomy_term_info("basihyal bone")
gene_info("socs5")

}
