% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-weights.R
\name{corpus_size}
\alias{corpus_size}
\title{Obtain the size of different corpora}
\usage{
corpus_size(
  corpus = c("taxon-annotations", "taxon-variation", "annotated-taxa",
    "gene-annotations", "genes", "states")
)
}
\arguments{
\item{corpus}{the name of the corpus, see \code{\link[=term_freqs]{term_freqs()}} for allowed values.}
}
\value{
The total size of the specified corpus as an integer number.
}
\description{
Obtains the size of a certain number of predefined corpora. The total size
of a corpus is important for calculating term frequencies. That is, for a given
corpus, the possible range for any term frequency is between 0 and the corpus size.
}
\details{
Corpus sizes are cached per session after they have first been obtained.
Thus, if the Phenoscape KB changes, a session needs to be restarted to
have those changes be reflected.
}
\examples{
corpus_size("taxon-variation")
corpus_size("annotated-taxa")
corpus_size("states")
corpus_size("genes")
}
