% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-weights.R
\name{corpus_size}
\alias{corpus_size}
\title{Obtain the size of different corpora}
\usage{
corpus_size(corpus = c("taxa", "states", "genes"))
}
\arguments{
\item{corpus}{the name of the corpus, currently one of "taxa", or "states".}
}
\value{
the size of the specified corpus as an integer number.
}
\description{
Obtains the size of a certain number of predefined corpora. The total size
of a corpus is important for calculating term frequencies.
}
\details{
Corpus sizes are cached per session after they have first been obtained.
Thus, if the Phenoscape KB changes, a session needs to be restarted to
have those changes be reflected.
}
\examples{
corpus_size("taxa")
corpus_size("states")
}
