% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_exclusivity.R
\name{mutual_exclusivity_pair_test}
\alias{mutual_exclusivity_pair_test}
\title{Determine whether two phenotypes are mutually exclusive.}
\usage{
mutual_exclusivity_pair_test(
  phenotype.a,
  phenotype.b,
  studies = NULL,
  charstates = NULL
)
}
\arguments{
\item{phenotype.a}{character, phenotype ID (IRI) or a phenotype object.}

\item{phenotype.b}{character, phenotype ID (IRI) or a phenotype object.}

\item{studies}{character, a vector of study IDs.}

\item{charstates}{dataframe, a dataframe obtained by running
charstates(list(phenotype.a, phenotype.b))}
}
\value{
is_pair_mutually_exclusive character, the mutual exclusivity result of two phenotypes.
The output can be one of the following:
'inconclusive_evidence' : default if no evidence supporting or contradicting
mutual exclusivity can be found.

'strong_compatibility'  : if the two phenotypes share one or more character-state value(s),
they are mutually compatible (strong evidence).

'strong_exclusivity'    : if the two phenotypes share one or more characters,
but do not share any states, then they are mutually exclusive (strong evidence).

'weak_compatibility'    : if the two phenotypes do not share any characters,
but there are taxa exhibiting both of their respective character states,
then they are mutually compatible (weak evidence).

'weak_exclusivity'      : if the two phenotypes do not share any characters,
and there are no taxa exhibiting both of their respective character states,
then they are mutually exclusive (weak evidence).
}
\description{
The function expects two phenotypes, not a vector.
Studies can be provided in case one is interested in determining the mutual
exclusivity among the two phenotypes based on the evidence found in a
particular set of studies.
A charstates dataframe is generally not required. It exists to make
computation for the mutual_exclusivity_test() function faster since
it repeatedly calls mutual_exclusivity_pair_test().
}
\examples{
\dontrun{
# get all studies in the database
studies <- pk_get_study_list()
study <- studies$id[studies$label == 'Dillman et al. (2016)']

# get phenotypes
phenotypes <- get_phenotypes(study=study)
#phenotypes <- phenotypes$id  # slower
phenotypes <- as.phenotype(phenotypes, withTaxa=TRUE)  # much faster, recommended

# determine mutual exclusivity between the first and fith phenotypes in the study
print(mutual_exclusivity_pair_test(phenotypes[1], phenotypes[5]))
}

}
