% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_exclusivity.R
\name{mutual_exclusivity_test}
\alias{mutual_exclusivity_test}
\title{Determine mutual exclusivity among multiple phenotypes.}
\usage{
mutual_exclusivity_test(phenotypes, studies = NULL, progress_bar = TRUE)
}
\arguments{
\item{phenotypes}{phenotype, a vector of phenotype objects or a vector of IRIs of
phenotypes for which one wants to determine the mutual exclusivity.
Passing phenotypes as phenotype objects results in much faster computation
than passing IRIs.}

\item{studies}{character, a vector of study IDs if one is interested to
determine mutual exclusivity based on the evidence from a particular set
of studies.}

\item{progress_bar}{logical, whether to print progress of the function.
If more than several hundred phenotypes are passed, it is suggested to track
the progress of the function by setting this parameter to TRUE.}
}
\value{
a list of a matrix and a dataframe containing mutual exclusivity
results for the phenotypes.
The matrix is square and its size is equal to the number of phenotypes.
Each cell contains a value that informs about the mutual exclusivity
of the phenotypes corresponding to the row and the column of the cell.
The dataframe contains five columns: the labels and ids of each phenotype
in a pair, and the mutual exclusivity value of the pair.
}
\description{
This function computes mutual exclusion among all pairs of phenotypes
in the phenotypes vector. It does so by calling the mutual_exclusivity_pair_test()
function for each phenotype pair.
}
\examples{
\dontrun{
# get studies
studies <- pk_get_study_list()
study <- studies$id[studies$label == 'Dillman et al. (2016)']

# get phenotypes
phenotypes <- get_phenotypes(study=study)
#phenotypes <- phenotypes$id  # slower
phenotypes <- as.phenotype(phenotypes, withTaxa=TRUE)  # faster, recommended

# compute mutual exclusivity
me <- mutual_exclusivity_test(phenotypes, study, progress_bar=TRUE)

# look at the resultant matrix
me$matrix

# look at resultant dataframe
head(me$dataframe[, c('label.1', 'label.2', 'mutual_exclusivity')])
}
}
