% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_classification.R
\name{pk_class}
\alias{pk_anatomical_class}
\alias{pk_class}
\alias{pk_phenotype_class}
\alias{pk_taxon_class}
\title{Classificatoin}
\usage{
pk_taxon_class(x, verbose = TRUE)

pk_anatomical_class(x, verbose = TRUE)

pk_phenotype_class(x, verbose = TRUE)
}
\arguments{
\item{x}{characters, name of the term}
}
\value{
A list containing data.frame
}
\description{
Return direct superclasses, direct subclasses, and equivalent classes of a given term
}

