% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_is_descendant.R
\name{pk_desc}
\alias{pk_desc}
\alias{pk_is_ancestor}
\alias{pk_is_descendant}
\title{pk_is_descendant}
\usage{
pk_is_descendant(taxon, candidates)

pk_is_ancestor(taxon, candidates)
}
\arguments{
\item{taxon}{}

\item{candidates,}{a list or vector of taxon IRIs, now only works for one candiate input}
}
\value{
a list of matching ids
}
\description{
pk_is_descendant
}

