% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_is_descendant.R
\name{pk_desc}
\alias{pk_desc}
\alias{pk_is_ancestor}
\alias{pk_is_descendant}
\title{pk_is_descendant}
\usage{
pk_is_descendant(taxon, candidates)

pk_is_ancestor(taxon, candidates)
}
\arguments{
\item{taxon}{character. The taxon name}

\item{candidates}{a list of taxa}
}
\value{
a list of TRUE/FALSE, indicating if the corresponding taxon in the candidate list is descendant/ancestor or not.
}
\description{
pk_is_descendant
}
\examples{
\dontrun{
pk_is_descendant("Halecostomi", c("Halecostomi", "Icteria", "Sciaenidae"))
pk_is_ancestor("Sciaenidae", c("Halecostomi", "Abeomelomys", "Sciaenidae"))
}
}

