% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_get_IRI.R
\name{pk_get_iri}
\alias{pk_get_iri}
\title{Resolve a text term into IRI}
\usage{
pk_get_iri(text, as, verbose = FALSE)
}
\arguments{
\item{text}{character. The term to be resolved}

\item{as}{character. Ontology type. For a taxon, use "vto", anatomical structure, "uberon", phenotype, "pato".}

\item{verbose}{logical: optional. If TRUE (default), informative messages printed.}
}
\value{
character. The resolved IRI.
}
\description{
Resolve a text term into IRI
}

