% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_study.R
\name{pk_get_study_list}
\alias{pk_get_study_list}
\title{Query the list of studies by taxa and anatomical entities}
\usage{
pk_get_study_list(taxon, entity, relation = "part of")
}
\arguments{
\item{taxon}{character. The taxon name.}

\item{entity}{character. The entity name.}

\item{relation}{character. Can be chosen from "part of" and "develops from".}
}
\value{
data.frame
}
\description{
Return studies containing taxa which are members of the optional input taxon
expression and are have annotated phenotypes which are relevant to the optional
input entity expression.
}
\examples{

slist <- pk_get_study_list(taxon = "Ictalurus", entity = "fin")
nex_list <- pk_get_study_xml(slist$id)
pk_get_study(nex_list)
pk_get_study_meta(nex_list)

}

