% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_terms.R
\name{pk_is_extinct}
\alias{pk_is_extinct}
\title{Test if a taxon is extinct.}
\usage{
pk_is_extinct(taxon)
}
\arguments{
\item{taxon}{character, the taxon name to be tested.}
}
\value{
logical, TRUE if extinct, FALSE if not
}
\description{
Test if a taxon is extinct.
}

