% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_ontotrace.R
\name{pk_ontotrace}
\alias{pk_ontotrace}
\title{Returns a NeXML-format evolutionary character matrix.}
\usage{
pk_ontotrace(..., relation = "part of", variable_only = TRUE)
}
\arguments{
\item{relation}{c("part of", "develops from")}

\item{variable_only}{logical}

\item{taxon}{characters}

\item{entity}{characters; anatomical class expression}
}
\value{
data.frame
}
\description{
Generate matrix of inferred presence/absence associations for anatomical structures
subsumed by the provided entity class expression, for any taxa within the provided
taxon class expression.
}

