% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_ontotrace.R
\name{pk_ontotrace}
\alias{pk_ontotrace}
\alias{pk_get_ontotrace}
\alias{pk_get_ontotrace_meta}
\title{Returns a NeXML-format evolutionary character matrix.}
\usage{
pk_get_ontotrace(nex)

pk_get_ontotrace_meta(nex)
}
\arguments{
\item{nex, }{a nexml object}
}
\value{
data.frame: The OntoTrace matrix.
}
\description{
Generate matrix of inferred presence/absence associations for anatomical structures
subsumed by the provided entity class expression, for any taxa within the provided
taxon class expression.
}
\examples{
\dontrun{
nex0 <- get_ontotrace_data(taxon = "Ictalurus", entity = "fin")

nex <- get_ontotrace_data(taxon = c("Ictalurus", "Ameiurus"), entity = "fin spine")
pk_get_ontotrace(nex)
pk_get_ontotrace_meta(nex)
}
}
