% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_ontotrace.R
\name{pk_ontotrace}
\alias{pk_ontotrace}
\title{Returns a NeXML-format evolutionary character matrix.}
\usage{
pk_ontotrace(taxon, entity, relation = "part of", get_metadata = TRUE,
  variable_only = TRUE)
}
\arguments{
\item{taxon}{character: Required. A single character string or a vector of taxa.}

\item{entity}{characters: Required. A single character string or a vector of anatomical class expressions.}

\item{relation}{character string: Optional. Has to be either "part of" or "develops from". Default is "part of".}

\item{get_metadata}{logical: Optional. If TRUE the result will contain the metadata data.frame. Default is TRUE}

\item{variable_only}{logical: Optional. Default is TRUE.}
}
\value{
data.frame: The OntoTrace matrix.
}
\description{
Generate matrix of inferred presence/absence associations for anatomical structures
subsumed by the provided entity class expression, for any taxa within the provided
taxon class expression.
}
\examples{
pk_ontotrace(taxon = "Ictalurus", entity = "fin")
pk_ontotrace(taxon = c("Ictalurus", "Ameiurus"), entity = "fin")
pk_ontotrace(taxon = c("Ictalurus", "Ameiurus"), entity = c("fin", "spine"), relation = "develops from")
}

