% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_terms.R
\name{pk_terms}
\alias{pk_anatomical_detail}
\alias{pk_phenotype_detail}
\alias{pk_taxon_detail}
\alias{pk_terms}
\title{Get traits}
\usage{
pk_taxon_detail(term, verbose = TRUE)

pk_anatomical_detail(term, verbose = TRUE)

pk_phenotype_detail(term, verbose = TRUE)
}
\arguments{
\item{term}{characters}

\item{verbose}{logical; If TRUE (default), informative messages printed.}
}
\value{
A data.frame with term id, label, and definition
}
\description{
To be expanded
}

