% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexml.R
\name{state_symbols2labels}
\alias{state_symbols2labels}
\title{Translates state symbols to labels in a character matrix}
\usage{
state_symbols2labels(nex, charmat, metacolumns = c(1, 2))
}
\arguments{
\item{nex}{the \link[RNeXML:nexml-class]{nexml} object from which the character matrix was obtained}

\item{charmat}{the character matrix (as a data.frame, usually obtained using \code{\link[=get_char_matrix]{get_char_matrix()}})}

\item{metacolumns}{the indexes of the columns in the matrix that contain metadata (such as taxa,
identifiers, etc), not character states. If not all of these columns are before all character
columns, then the returned matrix will move all of them to the front. The default is the first
two columns.}
}
\description{
Translates a character matrix using symbols for states to one using labels
instead. The character matrix must have been obtained from a \link[RNeXML:nexml-class]{nexml}
object, usually using \code{\link[=get_char_matrix]{get_char_matrix()}}, and the nexml object is required
for the translation.
}
\note{
If the nexml object contains multiple character matrices, currently the character labels
must be distinct between them.
}
\examples{
# obtain a (synthetic) Ontotrace matrix:
nex <- get_ontotrace_data(taxon = c("Ictalurus", "Ameiurus"), entity = "fin spine")
m <- get_char_matrix(nex, otus_id = FALSE)
# by default it uses symbols
m
# translate symbols to labels without having to re-obtain the matrix
state_symbols2labels(nex, m)
# if we obtained the matrix with otus_id, then we have one more metadata column
m <- get_char_matrix(nex, otus_id = TRUE)
state_symbols2labels(nex, m, metacolumns = c(1,2,3))

}
